fixes #295: Tasks name excessive when alot of files selected to be removed in one go
The new behaviour is to list only the file name bit (ignoring the full path), and listing all only if there are less than 5 entries. After the task contains 5 or more targets, we will only use the file names of the first, the median and the last element, followed by the number of other targets not included in the summary.